src/intent/requirements.txt
pandas==1.1.5
seaborn==0.11.2
networkx==2.2
pyyaml==5.4.1
jupyter==1.0.0
black==19.10b0
nltk==3.6.5
spacy==2.2.0
scikit-learn==0.24.2               
stanza==1.2.3    
allennlp==2.5.0                 
allennlp-models==2.5.0
mlflow==1.20.2
openpyxl==3.0.9
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
scikit-network==0.22.0
regex==2021.8.3
pigeon-jupyter==0.1.0
pywsd==1.2.4
wn==0.0.23
nltk_tgrep==1.0.6